Add project edit functionality and refactor project forms for consistency
